have just extraRpcs in extraRpcs.js and add rpcs.js with all
